Added package notificationengine that represents the architecture that sends the notifications by different STRATEGIES

+++ src/main/java/dev/alvartaco/notifications/repository/IMessageRepository.java
@@ -13,11 +13,8 @@ import java.util.Optional;
 public interface IMessageRepository {
 
         List<Message> findAllByMessageIdDesc() throws MessageException;
-
-        Optional<Message> findByMessageId(Long messageId) throws MessageException;
-
+        Optional<Message> findByMessageId(Integer messageId) throws MessageException;
         void create(@Valid Message message) throws MessageException;
-
         int count() throws MessageException;
 
 }
